Normalize style1.css to use 4-space indentation
This fixes two problems:

First, when we authored commit 7282162 ("Fix code style"), we had some trouble copying-and-pasting, and certain CSS rules ended up with the wrong amount of indentation.

Second, it looks like the original copy of style1.css, from https://github.com/UMM-CSci-3601/3601-lab1_HTML-CSS, uses a mix of 4-space and 2-space indentation. (Would it be appropriate to open an issue for this? It *is* fairly minor. I'll probably ask on Slack.)

I've changed everything to 4-space indentation because that seemed more prevalent in the original. If someone would rather use 2 spaces, though, I'd be fine with that! 😊