Enhance alignment of divs
Previously, the .pslo-courses-container div was being displayed slightly
wider than the .faculty-container and .tools-container divs on large
screens where the media query was in effect. .pslo-courses-container was
declaring a max-width of 53em, while .section (*not* .faculty-container
or .tools-container) was declaring a max-width of 48em, as well as 20px
worth of padding. 20px + 48em is about equal to 53em, but not quite, at
least on my screen.

This commit changes a few things.

First, we declare all widths on .something-container classes, rather
than on the .section class. This means that, when working with widths,
we don't have to worry about that extra 20px of padding; that padding is
on the interior, now.

Second, we rephrase widths in terms of percentages, rather than ems. On
large screens, all .something-container divs take up 75% of the width
of the body; on small screens, they take up 90%. (As part of this
change, we switched the `max-width` properties into `width` properties.)
This has a few advantages:

 - All .something-container divs will now stretch to fill the same
   width, even if their content is very short. (Previously, divs with
   content less wide than 48em wouldn't stretch to 48em.)
 - On very wide (or very zoomed-out) screens, the .something-container
   classes will take up exactly the same amount of screen space.
   (Previously, when zooming out, the content part of the screen would
   become very small and the margins on either side would become very
   large.)