Fix validation issues
Previously, we had two validation issues.

FIrst, the link to our google fonts contained the character "|", which we needed to replace with "%7C".

Second, one of our CSS rules used the property padding-block-end, which is still experimental. (It's only in the draft of the latest CSS standard.) We fixed this problem by removing that rule entirely.