Fix double underlines on certain browsers
There were two problems that made double-underlines fail to display on
Safari, Edge, and IE (including the mobile versions of Safari and
Edge.)

First, Safari does not support the CSS text-decoration property in its
shorthand form. (It wants text-decoration-line, text-decoration-color,
and text-decoration-style to be specified individually.) But listing the
individual properties also presents issues, since Edge and IE *only*
support the shorthand form--the exact opposite of what Safari supports.
Thankfully, Safari does recognize the vendor-prefixed
-webkit-text-decoration; we added this.

The second problem was that we were defining text-decoration on li
elements, some of which contained a elements. According to the spec,
however, text-decoration is not inherited by an element's children, so,
strictly speaking, the links never should have been underlined.

(Different browsers reacted to this situation differently; Chrome and
Firefox happily put a double-underline across the whole li element, but
Safari, Edge, and IE put a double underline across the ::after
pseudo-element and--for some reason--a single underline across the a
element.)

To fix the second problem, we changed the selectors so that the
text-decoration applies to #tools li.testing>a in addition
to #tools li.testing.